milestone_2/requirements.txt
pandas
requests
python-dotenv
newsdataapi
google-generativeai
nltk
wordcloud
matplotlib